+++ deps/media-playback/media-playback/media.c
@@ -373,7 +373,7 @@ void mp_media_next_audio(mp_media_t *m)
 		audio.data[i] = f->data[i];
 
 	audio.samples_per_sec = f->sample_rate * m->speed / 100;
-	audio.speakers = convert_speaker_layout(channels);
+	audio.speakers = convert_speaker_layout((uint8_t)channels);
 	audio.format = convert_sample_format(f->format);
 	audio.frames = f->nb_samples;
 	audio.timestamp = m->full_decode
